+++ packages/wonder-blocks-banner/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @khanacademy/wonder-blocks-banner
 
+## 3.1.1
+
+### Patch Changes
+
+-   Updated dependencies [16e1635a]
+    -   @khanacademy/wonder-blocks-icon-button@5.3.0
+
+## 3.1.0
+
+### Minor Changes
+
+-   9eecfc56: Adds optional `icon` prop for the `Banner` component. An icon asset (either imported from Phosphor or a imported custom icon) can be used. If not provided, default icons are used to represent the `kind` prop.
+
 ## 3.0.43
 
 ### Patch Changes